+++ api-base/src/commons/(providers)/example-base.provider.ts
@@ -5,12 +5,12 @@ import { ZanixProvider } from '@zanix/server'
  * @name ExampleBaseProvider
  * @description Provider class
  * @extends { ZanixProvider }
- * @summary Providers are for external connections.
+ * @summary Providers are for external APPs connections.
  * -------- This is for Adapters use only.
  */
 
 export class ExampleBaseProvider extends ZanixProvider {
   constructor() {
-    super('protocol://provider.com')
+    super('https://url.com')
   }
 }